% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{clusterUsingOntology}
\alias{clusterUsingOntology}
\title{OmicsON - private function.}
\usage{
clusterUsingOntology(chebiIdsDataFrame, rootColumnName,
  ontologyRepresentatnion)
}
\description{
OmicsON - private function.
}
