% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{createFunctionalInteractionsDataFrame}
\alias{createFunctionalInteractionsDataFrame}
\title{Create a functional interactions data frame.}
\usage{
createFunctionalInteractionsDataFrame(chebiToReactomeDataFrame,
  singleIdColumnName = "ontologyId", idsListColumnName = "ensembleIds",
  attachRootColumn = TRUE, rootIdCollumnName = "root")
}
\arguments{
\item{chebiToReactomeDataFrame}{A data frame returned from any of decorated functions.}

\item{singleIdColumnName}{Column with single ids per row.}

\item{idsListColumnName}{Column with multiple, list of ids per row.}

\item{attachRootColumn}{Use if you are interesting in attaching of root column into results.}

\item{rootIdCollumnName}{Name used in returned data frame for representation of root data column.}
}
\value{
Functional interactions data frame.
}
\description{
You can create a functional interactions data frame
from returned decorated data frames using this function.
All analytical functions works on functional interactions data frames
for more generic approach. In general this functions split mapping
one to many for bench of mappings one to one.
}
\examples{
ontology2EnsembleFunctionalInteractions <- OmicsON::createFunctionalInteractionsDataFrame(
chebiToReactomeDataFrame = decoratedByReactome,
singleIdColumnName = 'ontologyId', idsListColumnName = 'ensembleIds')

}
