% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPublicApi.R
\name{decorateByReactomeData}
\alias{decorateByReactomeData}
\title{Decorate data by data presented in Reactome database.}
\usage{
decorateByReactomeData(chebiMoleculesDf, chebiIdsColumnName,
  organismTaxonomyId = "9606")
}
\arguments{
\item{chebiMoleculesDf}{Data frame with column which represents ChEBI ids and set of data.}

\item{chebiIdsColumnName}{A string which represents a name of column with ChEBI ids.}

\item{organismTaxonomyId}{A string which represents an organism taxonomy id, for example human is '9606'.}
}
\value{
Data frame with columns:
- root - ChEBIs ids given by user,
- ontologyId - ChEBI ids used in the calculation, it is taken from ChEBI ontology base on root,
- ensembleIds - List including vector of Ensemble's Ids,
- uniProtIds - List including vector of UniProt's Ids,
- reactomeIds - List including vector of pathway's ids from Reactome DB,
- genesSymbolsFromEnsemble - List including vector of gen's symbols from Reactome DB base on pathway and Ensemble's Ids,
- genesSymbolsFromUniProt - List including vector of gen's symbols from Reactome DB base on pathway and UniProt's Ids,
}
\description{
Use this function on you data.frame with ChEBI ids. It decorates data
by data presented in Reactome database. It is done by searching of ontologically related
molecules presented in Reactome's pathways. If ChEBI id has representation in
Reactome pathways then we relay on it.
If ChEBI doesn't occure in Reactome mapping, then we are try to find representat
of that group by ontology in ChEBI database. There are also ids witch are not presented and do not have any no representants in ChEBI ontology,
so mapping and decoration is empty for that id.
}
\examples{
pathToFileWithLipidomicsData <- system.file(package = "OmicsON", "extdata", "nm-lipidomics.txt")
lipidomicsInputData <- read.table(pathToFileWithLipidomicsData, header = TRUE)
lipidomicsInputDataDecoratedByReactome <- OmicsON::decorateByReactomeData(
chebiMoleculesDf = lipidomicsInputData,
chebiIdsColumnName = "ChEBI",
organismTaxonomyId = '9606')

}
