% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPublicApi.R
\name{decorateByStringDbData}
\alias{decorateByStringDbData}
\title{Decorate data from previous step (Reactome decoration) by data presented in STRING DB database.}
\usage{
decorateByStringDbData(chebiIdsToReactomePathways,
  listOfEnsembleIdColumnName, stringOrganismId = "9606",
  stringDbVersion = "10", idsColumnName = "ontologyId",
  rootColumnName = "root")
}
\arguments{
\item{chebiIdsToReactomePathways}{Data frame returned from OmicsON::decorateByReactomeData().}

\item{listOfEnsembleIdColumnName}{A string which represents a name of column with Ensemble or UniProt ids.}

\item{stringOrganismId}{A string which represents an organism taxonomy id, for example human is '9606'.}

\item{stringDbVersion}{String with STRING DB version used to decoration.}

\item{idsColumnName}{Column name of ChEBI ids from ontology mapping.}

\item{rootColumnName}{Column name of original ChEBI ids.}
}
\value{
Input data frame with presented extra columns:
- stringIds - List including vector of all STRING's ids used in computations.
- stringGenesSymbolsExpand - List including vector of all neighbours find in STRING database.
- stringGenesSymbolsNarrow - List including vector of intersection of all neighbours per id from set of ids used in search.
}
\description{
In this part you search for any extra interactions of gens which you
find in Reactome. STRING calls them neighbours. To do it just put results achieved
from Reactome decoration step as input to this function.
}
\examples{
pathToFileWithLipidomicsData <- system.file(package = "OmicsON", "extdata", "nm-lipidomics.txt")
lipidomicsInputData <- read.table(pathToFileWithLipidomicsData, header = TRUE)
lipidomicsInputDataDecoratedByReactome <- OmicsON::decorateByReactomeData(
chebiMoleculesDf = lipidomicsInputData,
chebiIdsColumnName = "ChEBI",
organismTaxonomyId = '9606')
decoratedByStringBaseOnEnsembleIds <- OmicsON::decorateByStringDbData(
chebiIdsToReactomePathways = lipidomicsInputDataDecoratedByReactome,
listOfEnsembleIdColumnName = 'ensembleIds')
}
