% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChebiReactomeOntology.R
\name{firstExistsInReactomeChebiOntology}
\alias{firstExistsInReactomeChebiOntology}
\title{OmicsON - private function.}
\usage{
firstExistsInReactomeChebiOntology(baseData, rootColumnName)
}
\description{
OmicsON - private function.
}
