% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{getStringNeighbours}
\alias{getStringNeighbours}
\title{OmicsON - private function.}
\usage{
getStringNeighbours(chebiIdsToReactomePathways, stringOrganismId = 9606,
  stringDbVersion = "10", idsColumnName = "ontologyId",
  rootColumnName = "root", listOfEnsembleIdColumnName = "ensembleIds")
}
\description{
OmicsON - private function.
}
