% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{getSymbolsBaseOnEnsemblPeptidIdsUsingMyGenePackage}
\alias{getSymbolsBaseOnEnsemblPeptidIdsUsingMyGenePackage}
\title{OmicsON - private function.}
\usage{
getSymbolsBaseOnEnsemblPeptidIdsUsingMyGenePackage(gensIdsVector,
  organismTaxonomyId)
}
\description{
OmicsON - private function.
}
