% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{makeCanonicalCorrelationAnalysis}
\alias{makeCanonicalCorrelationAnalysis}
\title{Calculate CCA on data returned from decoration methods.}
\usage{
makeCanonicalCorrelationAnalysis(xNamesVector, yNamesVector, XDataFrame,
  YDataFrame, xCutoff = 1, yCutoff = 1, scalingFactor = 1)
}
\arguments{
\item{xNamesVector}{A vector of names from functional interactions data frame. Mostly genes symbols.}

\item{yNamesVector}{A vector of names from functional interactions data frame. Mostly root column.}

\item{XDataFrame}{A data frame with data to CCA analysis.
This argument is strongly connected with xNamesVector arg.}

\item{YDataFrame}{A data frame with data to CCA analysis.
This argument is strongly connected with yNamesVector arg.}

\item{xCutoff}{You can use it to remove highly correlated variables on XDataFrame.}

\item{yCutoff}{You can use it to remove highly correlated variables on YDataFrame.}

\item{scalingFactor}{scaling factor for input data.}
}
\value{
list with CCA analysis data.
}
\description{
Function is used to easily perform CCA analysis on functional interactions
data frames, those data frames are returned from OmicsON::createFunctionalInteractionsDataFrame
which works directly on data returned from functions responsible for decoration.
To close workflow, you can then use
dedicated plotting function on returned data, which is OmicsON::plotCanonicalCorrelationAnalysisResults.
}
\examples{
ontology2GenesSymboleFromEnsembleFunctionalInteractions <- OmicsON::createFunctionalInteractionsDataFrame(
decoratedByReactome,
singleIdColumnName = 'ontologyId',
idsListColumnName = 'genesSymbolsFromEnsemble')
ccaResultsEnsemble <- OmicsON::makeCanonicalCorrelationAnalysis(
xNamesVector = ontology2GenesSymboleFromEnsembleFunctionalInteractions$genesSymbolsFromEnsemble,
yNamesVector = ontology2GenesSymboleFromEnsembleFunctionalInteractions$root,
XDataFrame = transcriptomicsInputData,
YDataFrame = lipidomicsInputData, xCutoff = 0.5, yCutoff = 0.75)

}
