% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{makeCCAOnGroups}
\alias{makeCCAOnGroups}
\title{OmicsON - private function.}
\usage{
makeCCAOnGroups(groupsDefinitionDF, mappingDF,
  leftMappingColumnName = "root",
  rightMappingColumnName = "genesSymbolsFromEnsemble", groupsDataDF,
  mappingDataDF)
}
\description{
OmicsON - private function.
}
