% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{makePartialLeastSquaresRegression}
\alias{makePartialLeastSquaresRegression}
\title{Calculate PLS on data returned from decoration methods.}
\usage{
makePartialLeastSquaresRegression(xNamesVector, yNamesVector, XDataFrame,
  YDataFrame, formula = Y ~ X, ncompValue = 10, xCutoff = 1,
  yCutoff = 1, ...)
}
\arguments{
\item{xNamesVector}{A vector of names from functional interactions data frame. Mostly genes symbols.}

\item{yNamesVector}{A vector of names from functional interactions data frame. Mostly root column.}

\item{XDataFrame}{A data frame with data to PLS analysis.
This argument is strongly connected with xNamesVector arg.}

\item{YDataFrame}{A data frame with data to PLS analysis.
This argument is strongly connected with yNamesVector arg.}

\item{formula}{Formula to use in PLs, default is Y ~ X.}

\item{ncompValue}{Number of components in PLS analysis.}

\item{xCutoff}{You can use it to remove highly correlated variables on XDataFrame.}

\item{yCutoff}{You can use it to remove highly correlated variables on YDataFrame.}

\item{...}{Other arguments put into pls::plsr function, which is responsible for calculations.}
}
\value{
List with PLS results data.
}
\description{
Function is used to easily perform PLS analysis on functional interactions
data frames, those data frames are returned from OmicsON::createFunctionalInteractionsDataFrame
which works directly on data returned from functions responsible for decoration.
To close workflow, you can then use
dedicated plotting functions on returned data, which are
OmicsON::plotRmsepForPLS and OmicsON::plotRegression.
}
\examples{
ontology2GenesSymboleFromEnsembleFunctionalInteractions <- OmicsON::createFunctionalInteractionsDataFrame(
decoratedByReactome,
singleIdColumnName = 'ontologyId',
idsListColumnName = 'genesSymbolsFromEnsemble')
ccaResultsEnsemble <- OmicsON::makeCanonicalCorrelationAnalysis(
xNamesVector = ontology2GenesSymboleFromEnsembleFunctionalInteractions$genesSymbolsFromEnsemble,
yNamesVector = ontology2GenesSymboleFromEnsembleFunctionalInteractions$root,
XDataFrame = transcriptomicsInputData,
YDataFrame = lipidomicsInputData, xCutoff = 0.5, yCutoff = 0.75)

}
