% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{makePermutationTestOnCCA}
\alias{makePermutationTestOnCCA}
\title{OmicsON - private function.}
\usage{
makePermutationTestOnCCA(XDataFrame, YDataFrame, numberOfRowsForTestOnX,
  numberOfRowsForTestOnY, numberOfIterations = 100, countedCCA)
}
\description{
OmicsON - private function.
}
