% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{makePermutationTestOnPLS}
\alias{makePermutationTestOnPLS}
\title{OmicsON - private function.}
\usage{
makePermutationTestOnPLS(XDataFrame, YDataFrame, numberOfRowsForTestOnX,
  numberOfRowsForTestOnY, numberOfIterations = 100, countedPLS)
}
\description{
OmicsON - private function.
}
