% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{makePLSOnGroups}
\alias{makePLSOnGroups}
\title{OmicsON - private function.}
\usage{
makePLSOnGroups(groupsDefinitionDF, mappingDF,
  leftMappingColumnName = "root",
  rightMappingColumnName = "genesSymbolsFromEnsemble", groupsDataDF,
  mappingDataDF)
}
\description{
OmicsON - private function.
}
