% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{mapFromStringIdsToEnsembleIds}
\alias{mapFromStringIdsToEnsembleIds}
\title{OmicsON - private function.}
\usage{
mapFromStringIdsToEnsembleIds(vactofOfStringIds)
}
\description{
OmicsON - private function.
}
