% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{mapReactomePathwaysUnderOrganism}
\alias{mapReactomePathwaysUnderOrganism}
\title{OmicsON - private function.}
\usage{
mapReactomePathwaysUnderOrganism(chebiOntologyIds,
  organismTaxonomyId = "9606", idsColumnName = "ontologyId",
  rootColumnName = "root")
}
\description{
OmicsON - private function.
}
