% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChebiReactomeOntology.R
\name{mergeChEBIOntologyWithChildFavoring}
\alias{mergeChEBIOntologyWithChildFavoring}
\title{OmicsON - private function.}
\usage{
mergeChEBIOntologyWithChildFavoring(chebiOntologyDataFrame,
  rootColumnName = "root")
}
\description{
OmicsON - private function.
}
