% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{plotCanonicalCorrelationAnalysisResults}
\alias{plotCanonicalCorrelationAnalysisResults}
\title{Plot CCA analysis results.}
\usage{
plotCanonicalCorrelationAnalysisResults(ccaResults,
  x.name = "Transcriptomics", y.name = "Lipidomics", cv = 1,
  thirdLineText = "", ...)
}
\arguments{
\item{ccaResults}{Results from CCA analysis performed
by OmicsON::makeCanonicalCorrelationAnalysis function.}

\item{x.name}{Name for x-axis.}

\item{y.name}{Name for y-axis.}

\item{cv}{Number of canonical variates.}

\item{thirdLineText}{Extra text on plot.}

\item{...}{Arguments passed to helio.plot function.}
}
\description{
Function is used to plotting results from CCA analysis.
}
\examples{
OmicsON::plotCanonicalCorrelationAnalysisResults(ccaResults = ccaResults)

}
