% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapReactome.R
\name{plotRegression}
\alias{plotRegression}
\title{Plot PLS analysis results.}
\usage{
plotRegression(PLSResult, ncompValue = NULL)
}
\arguments{
\item{PLSResult}{Results from PLS analysis performed
by OmicsON::makePartialLeastSquaresRegression function.}

\item{ncompValue}{A number of components to print.}
}
\description{
Function is used to plotting results/predictions
from PLS analysis.
}
\examples{
OmicsON::plotRegression(PLSResults, ncompValue = 2)

}
