% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPublicApi.R
\name{setUpReactomeMapping}
\alias{setUpReactomeMapping}
\title{Set up space of mappings in package.}
\usage{
setUpReactomeMapping(ChEBI2ReactomeFileURL, Ensembl2ReactomeFileURL,
  UniProt2ReactomeFileURL)
}
\arguments{
\item{ChEBI2ReactomeFileURL}{A string which represents
full URL to mapping file on Reactome page. This file include ChEBI ids
mapped to Reactome pathway ids.}

\item{Ensembl2ReactomeFileURL}{A string which represents
full URL to mapping file on Reactome page. This file include Ensemble ids
mapped to Reactome pathway ids.}

\item{UniProt2ReactomeFileURL}{A string which represents
full URL to mapping file on Reactome page. This file include UniProt ids
mapped to Reactome pathway ids.}
}
\description{
Function is used to set up initial space of mappings in package.
It is necessary to set up this space before any analysis. It is crutial for the process
of mapping ChEBI, Ensemble and UniProt ids to Reactome pthways ids.
}
\examples{
OmicsON::setUpReactomeMapping(
ChEBI2ReactomeFileURL = "https://reactome.org/download/current/ChEBI2Reactome.txt",
Ensembl2ReactomeFileURL = "https://reactome.org/download/current/Ensembl2Reactome.txt",
UniProt2ReactomeFileURL = "https://reactome.org/download/current/UniProt2Reactome.txt")

}
